fix(tool): split pdf page number result card
