+++ tools/pdf-page-number-adder/client/preview-card.tsx
@@ -1,6 +1,5 @@
 import { useEffect, useMemo, useState } from "react"
 
-import { Badge } from "@workspace/ui/components/ui/badge"
 import { Button } from "@workspace/ui/components/ui/button"
 import {
   Card,
@@ -9,52 +8,30 @@ import {
   CardHeader,
   CardTitle,
 } from "@workspace/ui/components/ui/card"
-import {
-  Empty,
-  EmptyDescription,
-  EmptyHeader,
-  EmptyMedia,
-  EmptyTitle,
-} from "@workspace/ui/components/ui/empty"
-import {
-  ChevronLeft,
-  ChevronRight,
-  Download,
-  FileText,
-} from "@workspace/ui/icons"
+import { ChevronLeft, ChevronRight } from "@workspace/ui/icons"
 
 import { PreviewPage } from "./preview-page"
-import { useObjectUrl } from "./use-object-url"
 import { usePdfPagePreview } from "./use-pdf-page-preview"
-import { formatBytes, formatMessage } from "./utils"
+import { formatMessage } from "./utils"
 
-import type {
-  PageNumberFormOptions,
-  PdfPageNumberAdderMessages,
-  PdfPageNumberResult,
-} from "./types"
+import type { PageNumberFormOptions, PdfPageNumberAdderMessages } from "./types"
 
 type PreviewCardProps = Readonly<{
   file: File | null
-  isGenerating: boolean
   messages: PdfPageNumberAdderMessages
   options: PageNumberFormOptions
   pageCount: number
-  result: PdfPageNumberResult | null
   selectedPages: readonly number[]
 }>
 
 function PreviewCard({
   file,
-  isGenerating,
   messages,
   options,
   pageCount,
-  result,
   selectedPages,
 }: PreviewCardProps) {
   const [currentPreviewIndex, setCurrentPreviewIndex] = useState(0)
-  const resultUrl = useObjectUrl(result?.blob ?? null)
   const previewPageNumbers = useMemo(
     () => (selectedPages.length ? selectedPages : pageCount ? [1] : []),
     [pageCount, selectedPages]
@@ -98,7 +75,7 @@ function PreviewCard({
         <CardTitle>{messages.previewTitle}</CardTitle>
         <CardDescription>{messages.previewDescription}</CardDescription>
       </CardHeader>
-      <CardContent className="grid gap-6 lg:grid-cols-[minmax(0,1fr)_18rem]">
+      <CardContent className="flex min-w-0 flex-col gap-3">
         <div className="flex min-w-0 flex-col gap-3">
           <PreviewPage
             fontFamily={fontFamily}
@@ -128,12 +105,6 @@ function PreviewCard({
             {selectedPagesLabel}
           </p>
         </div>
-        <ResultPanel
-          isGenerating={isGenerating}
-          messages={messages}
-          result={result}
-          resultUrl={resultUrl}
-        />
       </CardContent>
     </Card>
   )
@@ -184,76 +155,6 @@ function PreviewNavigation({
       >
         <ChevronRight />
       </Button>
-    </div>
-  )
-}
-
-function ResultPanel({
-  isGenerating,
-  messages,
-  result,
-  resultUrl,
-}: Readonly<{
-  isGenerating: boolean
-  messages: PdfPageNumberAdderMessages
-  result: PdfPageNumberResult | null
-  resultUrl: string | null
-}>) {
-  return (
-    <div aria-live="polite" className="flex min-w-0 flex-col gap-4">
-      <div>
-        <h3 className="text-base font-medium">{messages.resultTitle}</h3>
-        <p className="mt-1 text-sm text-muted-foreground">
-          {messages.resultDescription}
-        </p>
-      </div>
-      {result ? (
-        <div className="flex flex-col gap-4 rounded-lg border bg-background p-4">
-          <div>
-            <h4 className="text-sm font-medium">{messages.resultReadyTitle}</h4>
-            <p className="mt-1 text-sm break-all text-muted-foreground">
-              {result.fileName}
-            </p>
-          </div>
-          <div className="flex flex-wrap gap-2">
-            <Badge variant="secondary">
-              {messages.pageCountLabel}: {result.pageCount}
-            </Badge>
-            <Badge variant="outline">
-              {messages.outputSizeLabel}: {formatBytes(result.blob.size)}
-            </Badge>
-          </div>
-          {resultUrl ? (
-            <Button
-              asChild
-              className="max-w-full text-center whitespace-normal"
-            >
-              <a download={result.fileName} href={resultUrl}>
-                <Download data-icon="inline-start" />
-                {messages.downloadPdfLabel}
-              </a>
-            </Button>
-          ) : null}
-        </div>
-      ) : (
-        <Empty className="min-h-48 border border-dashed border-border/80 bg-muted/20">
-          <EmptyHeader>
-            <EmptyMedia variant="icon">
-              <FileText />
-            </EmptyMedia>
-            <EmptyTitle>
-              {isGenerating
-                ? messages.generatingTitle
-                : messages.emptyResultTitle}
-            </EmptyTitle>
-            <EmptyDescription>
-              {isGenerating
-                ? messages.generatingDescription
-                : messages.emptyResultDescription}
-            </EmptyDescription>
-          </EmptyHeader>
-        </Empty>
-      )}
     </div>
   )
 }
